requirements.txt
azure-functions
vaderSentiment